Update utils.py
In this updated function (def plot_boxes_cv2), the confidence is formatted to have no decimal places when it's converted to a string in the msg variable (using :.0f in the formatted string). Also, if no savename is provided, the image will be displayed in a window. You can close this window by pressing any key. If you provide a savename, the image will be saved to the specified file instead of being displayed.